feat: pepero stick height function
