+++ CodeStarterCamp_Week1/main.swift
@@ -1,23 +1,16 @@
+//
+//  step3.swift
+//  CodeStarterCamp_Week1
+//
+//  Created by MacPro on 2022/04/18.
+//
+
 import Foundation
 
-func drawIcecream(forwidth width: Int, forheight height: Int, forStickHeight stickHeight: Int) {
+func makeStick(height: Int) {
     for _ in 1...height {
-        for _ in 1...width {
-            print("*", terminator: "")
-        }
-        print("")
-    }
-    
-    for _ in 1...stickHeight {
-        for _ in 1...((width-3)/2) {
-            print(" ", terminator: "")
-        }
-        print("| |", terminator: "")
-        for _ in 1...((width-3)/2) {
-            print(" ", terminator: "")
-        }
-        print("")
+        print(" | | ")
     }
 }
 
-drawIcecream(forwidth: 20, forheight: 10, forStickHeight: 5)
+makeStick(height: 5)
