--- Week1 Step2.drawio
@@ -1,1 +1,0 @@
-<mxfile host="app.diagrams.net" modified="2022-04-14T02:12:15.262Z" agent="5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/100.0.4896.60 Safari/537.36" etag="6YiH_6Ou5Cv8ANgVY8g4" version="17.4.4" type="device"><diagram id="C5RBs43oDa-KdzZeNtuy" name="Page-1">5Vpbd6M2EP41vOw5yZEAYXg0jpO2m+7lOG26fSNYsWkw8oKc2PvrK4FkJMCYrONL1i8+0jAISfPNzDeSDWswW96kwXz6Jxnj2DDBeGlYV4Zpeq7LfrlgVQgcq1cIJmk0LkSwFIyiH1gIgZAuojHONEVKSEyjuS4MSZLgkGqyIE3Ji672SGL9q/NggmuCURjEdel9NKbTQuqavVL+G44mU/ll6HjFk1kglcVKsmkwJi+KyBoa1iAlhBat2XKAY753cl/uf1/dx7dPzs0fX7PvwV/+x7tPf18Ug12/5pX1ElKc0Lcd2iyGfg7ihdgvsVa6khuYkkUyxnwQYFj+lM5i1oSs+R+mdCUMHiwoYSKS0imZkCSIbwmZC71HklChBnkfJ+M+NyzrP8QkfCpE11Eci2+wntB3WS+jKXla244PsDYEV46DBxz7Qfg0ySc6IDFJ2aOEJJgPNWZgEGspJzcspX7HvRU2yMgiDXGLnlXo8e8q+BMWucFkhmm6YgopjgMaPesgDQTWJ2u99atfSMTmZwLhlmtQCq+EPaAPQYN0gql4q0QFayjTKEU5Vl6BG6uOm+HAcH3DG/CGNzQ82IIkbsWXaUTxaB7k2/nCYo+OLhU1bBf8SRxkmbD5Fki8zqTPOKV42Wos+dTRN90V3ZcyrkBpl6kSU2yw2byaYVqsgD/f/Uj/HX396DGv+BLj0e3doIv3bncA3b8bPbjB0xVLNSL+51yqcZVi9wpEt+jt6mk7mQLUTBGFOExxMBMw7bOnLPyZTszW4D+krDXhLabGtWQO4mpuXSujUfik6ti7edcG6He1WmcXWcelbT7i7MtH7LPwEcmstvmIdUwfgQ2mcL4vOH/yP5RNwxzwpeB0FiUBxQXe1cdKq+ImPPf4Tp6EfKMPjD6qGXuLY+hedCg3seCx3QQ25ZJie3km1vZQ7j9/cJHlkM+jG5ovNePkJvk2HMmB2MSKsYSxfkE/tLrmqnZ8XIBLCCHUMHIh8fCz1FGqkMfHDO/KCptXdRrBVqWO6JSMftToW6fsH3KqfmV4Lg+ww77hA8HfWdHPhIWEs1wQKVSGhVbfMnybNxjj79v525L0F0O6oHzbuq5zlSmZPSyy7TzlMOHYQ5qn2XY9GpuoIRpDsK9w7HZwpLKMDnltFIWbti6vqZcR/Udpf+PtSyR6V0vl0dVKq6P2SFo2WEXZ9qZdRzvGwWXFeMLqyKsYs5h/rYTeWovXkvSGWpxZL1gpanOukG2esK1/xoGodVoVddsDr1GXiyhBXEz3bfNFnRHKmOQWUYUFJ5SHIhZqTBmoWJvJZdip8D3xykCJTu+BASIXHZsBOmeZvGHn7I2Omb1h84kbg7ebJ2Keuu08514b8sj+1DBvIfvUMO+dJ+ZRV8y7R8U8asP866irODtr4K5NZwjvl8pagLM61ct6oIHNyusvnc2ifRmyVzNkfiywE6OtOs2a4V56PU9jubZrtfJc3vmC04gtFqeH4r4bskl5AODatqdTsrod90iPEdQJSu/U6TGqMKoeaie8Vf1WxnvIUxSzzorPIim5HZOScvl+jKTUUJq/dyLmonpoOSwRMxvo7RlgXu7xdsybx8S8nGbjzQ1rBTMO4+Qhm+d90CTqDeRvJ/Vf4JKn6mXe8b2sXuJ/+nwWfmZ29bN2WgYuAZKbKK9lLHs3vnUAQmHWzK7+n2DTeb5aZ/nv/5Dfklec8q8JoMEf13+sUP2xynzfzh/r92X7q4vgCZ38b0pnSvVjuXoZa9bPivZY/ViV43AIqlc9XcsfG1a4lrmf8scB1Rl77fcD1RfkxI5f/zQd+7DQBHnMeReZHjLJ3lI965b/OS72vPzjtjX8Hw==</diagram></mxfile>
